CREP-23 Add dynamic profiles to pages directory. Fix routes/imports of related files
